HESML_Library/BERTExperiments/requirements.txt
absl-py==0.8.1
astor==0.8.0
bert==2.2.0
bert-serving-client==1.9.8
bert-serving-server==1.9.8
bert-tensorflow==1.0.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
erlastic==2.0.0
gast==0.2.2
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
GPUtil==1.4.0
grpcio==1.24.3
h5py==2.10.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
numpy==1.17.3
oauthlib==3.1.0
opt-einsum==3.1.0
pkg-resources==0.0.0
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pyzmq==18.1.0
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scipy==1.3.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
